moving position into its own file
