+++ src/inspection/visitNode.ts
@@ -3,9 +3,10 @@
 
 import { Node } from ".";
 import { ArrayUtils, CommonError, isNever, Option } from "../common";
-import { Token, TokenPosition } from "../lexer";
+import { TokenPosition } from "../lexer";
 import { Ast, NodeIdMap, ParserContext } from "../parser";
-import { Position, State } from "./inspection";
+import { State } from "./inspection";
+import { isPositionOnAstNode, isPositionOnXorNode, Position } from "./position";
 import { PositionIdentifierKind } from "./positionIdentifier";
 
 export function visitNode(state: State, xorNode: NodeIdMap.TXorNode): void {
@@ -522,178 +523,6 @@ function isParentOfNodeKind(
             throw isNever(parent);
     }
 }
-
-function isPositionBeforeXorNode(position: Position, xorNode: NodeIdMap.TXorNode): boolean {
-    switch (xorNode.kind) {
-        case NodeIdMap.XorNodeKind.Ast:
-            return isPositionBeforeAstNode(position, xorNode.node);
-
-        case NodeIdMap.XorNodeKind.Context:
-            return true;
-
-        default:
-            throw isNever(xorNode);
-    }
-}
-
-function isPositionOnXorNode(position: Position, xorNode: NodeIdMap.TXorNode): boolean {
-    switch (xorNode.kind) {
-        case NodeIdMap.XorNodeKind.Ast:
-            return isPositionOnAstNode(position, xorNode.node);
-
-        case NodeIdMap.XorNodeKind.Context:
-            return true;
-
-        default:
-            throw isNever(xorNode);
-    }
-}
-
-function isPositionAfterXorNode(position: Position, xorNode: NodeIdMap.TXorNode): boolean {
-    switch (xorNode.kind) {
-        case NodeIdMap.XorNodeKind.Ast:
-            return isPositionAfterAstNode(position, xorNode.node);
-
-        case NodeIdMap.XorNodeKind.Context:
-            return true;
-
-        default:
-            throw isNever(xorNode);
-    }
-}
-
-function isPositionBeforeContextNode(position: Position, contextNode: ParserContext.Node): boolean {
-    const maybeTokenStart: Option<Token> = contextNode.maybeTokenStart;
-    if (maybeTokenStart === undefined) {
-        return false;
-    }
-    const tokenStart: Token = maybeTokenStart;
-
-    return isPositionBeforeTokenPosition(position, tokenStart.positionStart);
-}
-
-function isPositionOnContextNode(
-    nodeIdMapCollection: NodeIdMap.Collection,
-    position: Position,
-    contextNode: ParserContext.Node,
-): boolean {
-    return (
-        !isPositionBeforeContextNode(position, contextNode) &&
-        !isPositionAfterContextNode(nodeIdMapCollection, position, contextNode)
-    );
-}
-
-function isPositionAfterContextNode(
-    nodeIdMapCollection: NodeIdMap.Collection,
-    position: Position,
-    contextNode: ParserContext.Node,
-): boolean {
-    const maybeRightMostAstNode: Option<Ast.TNode> = NodeIdMap.maybeRightMostAstDescendant(
-        nodeIdMapCollection,
-        contextNode.id,
-    );
-    if (maybeRightMostAstNode === undefined) {
-        return false;
-    }
-    const rightMostAstNode: Ast.TNode = maybeRightMostAstNode;
-
-    return isPositionAfterAstNode(position, rightMostAstNode);
-}
-
-function isPositionBeforeAstNode(position: Position, astNode: Ast.TNode): boolean {
-    return isPositionBeforeTokenPosition(position, astNode.tokenRange.positionStart);
-}
-
-function isPositionOnAstNode(position: Position, astNode: Ast.TNode): boolean {
-    return !isPositionBeforeAstNode(position, astNode) && !isPositionAfterAstNode(position, astNode);
-}
-
-function isPositionAfterAstNode(position: Position, astNode: Ast.TNode): boolean {
-    return isPositionAfterTokenPosition(position, astNode.tokenRange.positionEnd);
-}
-
-function isPositionBeforeTokenPosition(position: Position, tokenPositionStart: TokenPosition): boolean {
-    const positionLineNumber: number = position.lineNumber;
-
-    if (positionLineNumber < tokenPositionStart.lineNumber) {
-        return true;
-    } else if (positionLineNumber > tokenPositionStart.lineNumber) {
-        return false;
-    } else {
-        return position.lineCodeUnit <= tokenPositionStart.codeUnit;
-    }
-}
-
-function isPositionAfterTokenPosition(position: Position, tokenPositionEnd: TokenPosition): boolean {
-    const positionLineNumber: number = position.lineNumber;
-
-    if (positionLineNumber < tokenPositionEnd.lineNumber) {
-        return false;
-    } else if (positionLineNumber > tokenPositionEnd.lineNumber) {
-        return true;
-    } else {
-        return position.lineCodeUnit > tokenPositionEnd.codeUnit;
-    }
-}
-
-// function isTokenPositionOnOrBeforeBeforePostion(tokenPosition: TokenPosition, position: Position): boolean {
-//     return (
-//         isPositionOnTokenPosition(tokenPosition, position) || isPositionOnOrAfterTokenPosition(tokenPosition, position)
-//     );
-// }
-
-// function isPositionOnTokenPosition(tokenPosition: TokenPosition, position: Position): boolean {
-//     return tokenPosition.lineNumber === position.lineNumber && tokenPosition.lineCodeUnit - 1 === position.lineCodeUnit;
-// }
-
-// function isPositionOnOrAfterTokenPosition(tokenPosition: TokenPosition, position: Position): boolean {
-//     return (
-//         tokenPosition.lineNumber < position.lineNumber ||
-//         (tokenPosition.lineNumber === position.lineNumber && tokenPosition.lineCodeUnit - 1 < position.lineCodeUnit)
-//     );
-// }
-
-// function isPositionInXorNode(position: Position, xorNode: NodeIdMap.TXorNode): boolean {
-//     switch (xorNode.kind) {
-//         case NodeIdMap.XorNodeKind.Ast:
-//             return isPositionInTokenRange(position, xorNode.node.tokenRange);
-
-//         case NodeIdMap.XorNodeKind.Context:
-//             return true;
-
-//         default:
-//             throw isNever(xorNode);
-//     }
-// }
-
-// function isPositionInTokenRange(position: Position, tokenRange: Ast.TokenRange): boolean {
-//     const tokenRangeStart: TokenPosition = tokenRange.positionStart;
-//     const offsetPositionLineCodeUnit: number = position.lineCodeUnit - 1;
-
-//     if (
-//         position.lineNumber > tokenRangeStart.lineNumber ||
-//         (position.lineNumber === tokenRangeStart.lineNumber &&
-//             offsetPositionLineCodeUnit < tokenRangeStart.lineCodeUnit)
-//     ) {
-//         return false;
-//     }
-
-//     const tokenRangeEnd: TokenPosition = tokenRange.positionEnd;
-//     if (
-//         position.lineNumber < tokenRangeEnd.lineNumber ||
-//         (position.lineNumber === tokenRangeEnd.lineNumber && offsetPositionLineCodeUnit > tokenRangeEnd.lineCodeUnit)
-//     ) {
-//         return false;
-//     }
-
-//     return true;
-// }
-
-// function isTokenPositionOnOrBeforeBeforePostion(tokenPosition: TokenPosition, position: Position): boolean {
-//     return (
-//         isPositionOnTokenPosition(tokenPosition, position) || isPositionOnOrAfterTokenPosition(tokenPosition, position)
-//     );
-// }
 
 function addToScopeIfNew(state: State, key: string, xorNode: NodeIdMap.TXorNode): void {
     const scopeMap: Map<string, NodeIdMap.TXorNode> = state.result.scope;
